backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
tzdata>=2024.2
pytest>=8.0.0
requests>=2.31.0
python-multipart>=0.0.9
typer>=0.9.0
alembic>=1.13.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
apscheduler==3.10.4
pytz==2024.1
aiofiles==23.2.1
